+++ pimd/pim_zebra.h
@@ -23,7 +23,6 @@
 #include <zebra.h>
 #include "zclient.h"
 
-#include "pim_igmp.h"
 #include "pim_ifchannel.h"
 
 void pim_zebra_init(void);
@@ -31,15 +30,6 @@ void pim_zebra_zclient_update(struct vty *vty);
 
 void pim_scan_individual_oil(struct channel_oil *c_oil, int in_vif_index);
 void pim_scan_oil(struct pim_instance *pim_matcher);
-
-void igmp_anysource_forward_start(struct pim_instance *pim,
-				  struct gm_group *group);
-void igmp_anysource_forward_stop(struct gm_group *group);
-
-void igmp_source_forward_start(struct pim_instance *pim,
-			       struct gm_source *source);
-void igmp_source_forward_stop(struct gm_source *source);
-void igmp_source_forward_reevaluate_all(struct pim_instance *pim);
 
 void pim_forward_start(struct pim_ifchannel *ch);
 void pim_forward_stop(struct pim_ifchannel *ch);
